test: relocate declarative state/temp-file tests into the existing suites
